Revert "Auth use case 3/4"
